Make stubbing generic and add to the httpclient version as well
