+++ async-http-client-backend/zio/src/main/scala/sttp/client/asynchttpclient/zio/package.scala
@@ -4,11 +4,8 @@ import _root_.zio._
 import _root_.zio.stream._
 import sttp.client._
 import sttp.client.asynchttpclient.zio.SttpClientStubbing.StubbingWhenRequest
-import sttp.client.impl.zio.RIOMonadAsyncError
-import sttp.client.monad.MonadError
-import sttp.client.testing.{SttpBackendStub, WebSocketStub}
+import sttp.client.impl.zio.SttpClientStubbingBase
 import sttp.client.ws.{WebSocket, WebSocketResponse}
-import sttp.model.{Headers, StatusCode}
 
 package object zio {
 
@@ -55,130 +52,9 @@ package object zio {
       ZioWebSocketHandler().flatMap(handler => ZIO.accessM(env => env.get[Service].openWebsocket(request, handler)))
   }
 
-  object SttpClientStubbing {
-
-    trait Service {
-      def whenRequestMatchesPartial(
-          partial: PartialFunction[Request[_, _], Response[_]]
-      ): URIO[SttpClientStubbing, Unit]
-
-      private[zio] def update(
-          stub: SttpBackendStub[Task, Stream[Throwable, Byte], WebSocketHandler] => SttpBackendStub[
-            Task,
-            Stream[Throwable, Byte],
-            WebSocketHandler
-          ]
-      ): UIO[Unit]
-    }
-
-    private class StubWrapper(stub: Ref[SttpBackendStub[Task, Stream[Throwable, Byte], WebSocketHandler]])
-        extends Service {
-      override def whenRequestMatchesPartial(
-          partial: PartialFunction[Request[_, _], Response[_]]
-      ): URIO[SttpClientStubbing, Unit] =
-        update(_.whenRequestMatchesPartial(partial))
-
-      override private[zio] def update(
-          f: SttpBackendStub[Task, Stream[Throwable, Byte], WebSocketHandler] => SttpBackendStub[
-            Task,
-            Stream[Throwable, Byte],
-            WebSocketHandler
-          ]
-      ) = stub.update(f)
-    }
-
-    final case class StubbingWhenRequest private[zio] (p: Request[_, _] => Boolean) {
-      val thenRespondOk: URIO[SttpClientStubbing, Unit] =
-        thenRespondWithCode(StatusCode.Ok)
-
-      def thenRespondNotFound(): URIO[SttpClientStubbing, Unit] =
-        thenRespondWithCode(StatusCode.NotFound, "Not found")
-
-      def thenRespondServerError(): URIO[SttpClientStubbing, Unit] =
-        thenRespondWithCode(StatusCode.InternalServerError, "Internal server error")
-
-      def thenRespondWithCode(status: StatusCode, msg: String = ""): URIO[SttpClientStubbing, Unit] = {
-        thenRespond(Response(msg, status, msg))
-      }
-
-      def thenRespond[T](body: T): URIO[SttpClientStubbing, Unit] =
-        thenRespond(Response[T](body, StatusCode.Ok, "OK"))
-
-      def thenRespond[T](resp: => Response[T]): URIO[SttpClientStubbing, Unit] =
-        URIO.accessM(_.get.update(_.whenRequestMatches(p).thenRespond(resp)))
-
-      def thenRespondCyclic[T](bodies: T*): URIO[SttpClientStubbing, Unit] = {
-        thenRespondCyclicResponses(bodies.map(body => Response[T](body, StatusCode.Ok, "OK")): _*)
-      }
-
-      def thenRespondCyclicResponses[T](responses: Response[T]*): URIO[SttpClientStubbing, Unit] = {
-        for {
-          q <- Queue.bounded[Response[T]](responses.length)
-          _ <- q.offerAll(responses)
-          _ <- thenRespondWrapped(q.take.flatMap(t => q.offer(t) as t))
-        } yield ()
-      }
-
-      def thenRespondWrapped(resp: => Task[Response[_]]): URIO[SttpClientStubbing, Unit] = {
-        val m: PartialFunction[Request[_, _], Task[Response[_]]] = {
-          case r if p(r) => resp
-        }
-        URIO.accessM(_.get.update(_.whenRequestMatches(p).thenRespondWrapped(m)))
-      }
-
-      def thenRespondWrapped(resp: Request[_, _] => Task[Response[_]]): URIO[SttpClientStubbing, Unit] = {
-        val m: PartialFunction[Request[_, _], Task[Response[_]]] = {
-          case r if p(r) => resp(r)
-        }
-        URIO.accessM(_.get.update(_.whenRequestMatches(p).thenRespondWrapped(m)))
-      }
-
-      def thenRespondWebSocket[WS_RESULT](result: WS_RESULT): URIO[SttpClientStubbing, Unit] =
-        thenRespondWebSocket(Headers(List.empty), result)
-
-      def thenRespondWebSocket[WS_RESULT](headers: Headers, result: WS_RESULT): URIO[SttpClientStubbing, Unit] =
-        URIO.accessM(_.get.update(_.whenRequestMatches(p).thenRespondWebSocket(headers, result)))
-
-      def thenRespondWebSocket(wsStub: WebSocketStub[_]): URIO[SttpClientStubbing, Unit] =
-        thenRespondWebSocket(Headers(List.empty), wsStub)
-
-      def thenRespondWebSocket(headers: Headers, wsStub: WebSocketStub[_]): URIO[SttpClientStubbing, Unit] =
-        URIO.accessM(_.get.update(_.whenRequestMatches(p).thenRespondWebSocket(headers, wsStub)))
-
-      def thenHandleOpenWebSocket[WS_RESULT](
-          useHandler: WebSocketHandler[WS_RESULT] => WS_RESULT
-      ): URIO[SttpClientStubbing, Unit] =
-        thenHandleOpenWebSocket(Headers(List.empty), useHandler)
-
-      def thenHandleOpenWebSocket[WS_RESULT](
-          headers: Headers,
-          useHandler: WebSocketHandler[WS_RESULT] => WS_RESULT
-      ): URIO[SttpClientStubbing, Unit] = {
-        URIO.accessM(_.get.update(_.whenRequestMatches(p).thenHandleOpenWebSocket(headers, useHandler)))
-      }
-
-    }
-
-    val layer: ZLayer[Any, Nothing, SttpClientStubbing with SttpClient] =
-      ZLayer.fromEffectMany(for {
-        stub <- Ref.make(AsyncHttpClientZioBackend.stub)
-        stubber = new StubWrapper(stub)
-        proxy = new SttpClient.Service {
-          override def send[T](request: Request[T, Stream[Throwable, Byte]]): Task[Response[T]] =
-            stub.get >>= (_.send(request))
-
-          override def openWebsocket[T, WS_RESULT](
-              request: Request[T, Stream[Throwable, Byte]],
-              handler: WebSocketHandler[WS_RESULT]
-          ): Task[WebSocketResponse[WS_RESULT]] =
-            stub.get >>= (_.openWebsocket(request, handler))
-
-          override def close(): Task[Unit] =
-            stub.get >>= (_.close())
-
-          override def responseMonad: MonadError[Task] = new RIOMonadAsyncError[Any]()
-        }
-      } yield Has.allOf[SttpClientStubbing.Service, SttpClient.Service](stubber, proxy))
+  object SttpClientStubbing extends SttpClientStubbingBase[Any, Stream[Throwable, Byte], WebSocketHandler] {
+    override private[sttp] def serviceTag: Tag[SttpClientStubbing.Service] = implicitly
+    override private[sttp] def sttpBackendTag: Tag[SttpClient.Service] = implicitly
   }
 
   object stubbing {
@@ -193,5 +69,4 @@ package object zio {
     ): URIO[SttpClientStubbing, Unit] =
       ZIO.accessM(_.get.whenRequestMatchesPartial(partial))
   }
-
 }
